app/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.2
djangorestframework==3.14.0
Levenshtein==0.21.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pytz==2023.3
rapidfuzz==3.0.0
sqlparse==0.4.4
typing-extensions==4.6.3
xmltodict==0.13.0
